pyproject.toml
[project]
name = "arbitraz"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "grpcio==1.60.1",
    "ibm-watsonx-ai>=1.4.2",
    "notebook>=7.4.7",
    "numpy",
    "pandas-stubs==2.3.2.250926",
    "python-dotenv>=1.1.1",
]
